% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_cinema.R
\name{adding_city}
\alias{adding_city}
\title{Update markers of cities}
\usage{
adding_city(mapID, x)
}
\arguments{
\item{x}{A \code{database} with at least the colums of Latitude, Longitude,
cinema, Address, Telephone and Webpage related to several movies.}

\item{mapId}{The ID of a map \code{plot} created with function
\code{plot_city} that has to be updated.}
}
\value{
The updated plot of function \code{plot_city} with the
markers specified by the filters in a shiny app.
}
\description{
This function updates the plot generated with function
\code{plot_city} with the new markers avaiable on a new data set, that's
created base in the reactive results of a shiny app.
}
\examples{
adding_city(mapID,x)devto
}
\author{
Ines Guardans, Guillaume Lakah, Monica Navarro & Mathieu Schnyder
}
