% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allmovies_df}
\alias{allmovies_df}
\title{Information of all movies for 3 days in SWitzerland.}
\format{A data frame with approximately 8677 rows and 42 variables:
\describe{
  \item{movie_title}{Name of the movie}
  \item{cinema}{Name of the cinema}
  \item{town}{Town where the cinema is located}
  \item{viewing_times}{Hour and minute when the movie will be displayed}
  \item{movie_date}{Date when the movie will be displayed}
  \item{genres}{Binary variables that indicated if the movie corresponds
  to that genre (if = 1)}
  \item{movie_rating}{Numeric variable from 1 to 5 that contains
  the rating of the movie}
  \item{movie_poster}{String with link for the movie poster}
  \item{Known as}{Alternative name of the cinema}
  \item{Address}{Adrdess of the cinema}
  \item{Webpage}{Webpage of the cinema when available}
  \item{Telephone}{Telephone of the cinema when available}
  \item{Latitude}{Latitude of the cinema}
  \item{Longitude}{Longitude of the cinema}
}}
\source{
Scrapped from \url{http://www.cineman.ch/}
}
\usage{
allmovies_df
}
\description{
A dataset containing the viewing times and other attributes of all movies
displayed in Switzerland for the next 3 days.
}
\keyword{datasets}
