% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genrelist}
\alias{genrelist}
\title{Information of all genres of movies for 3 days in SWitzerland.}
\format{A data frame with approximately 8677 rows and 42 variables:
\describe{
  \item{genrelist}{All genres in movies for the nest 3 days}
}}
\source{
Scrapped from \url{http://www.cineman.ch/}
}
\usage{
genrelist
}
\description{
A dataset containing all genres of all movies displayed in Switzerland for
the next 3 days.
}
\keyword{datasets}
