% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateMovieData.r
\name{updateMovieData}
\alias{updateMovieData}
\title{Update Movie Data}
\usage{
updateMovieData(updatemovies = TRUE)
}
\arguments{
\item{updatemovies}{A binary \code{string} with options TRUE or FALSE to scrap
the movie data of the following 5 days.Default value is \code{TRUE}.}
}
\value{
Datasets for all the genres available and for movies in Switzerland
for the following 5 days. Datasets are stored in ./data as .rda files.
}
\description{
Scraps 2 data sets from the webpage \url{https://www.cineman.ch}
that work directly with the shiny app loaded in the function
\code{runMovieApp}
}
\examples{
updateMovieData(updatemovies = TRUE)
}
\author{
Ines Guardans, Guillaume Lakah, Monica Navarro & Mathieu Schnyder
}
